+++ packages/webview_flutter/ios/Classes/FlutterWebView.m
@@ -32,6 +32,7 @@
   WKWebView* _webView;
   int64_t _viewId;
   FlutterMethodChannel* _channel;
+  NSString* _currentUrl;
 }
 
 - (instancetype)initWithFrame:(CGRect)frame
@@ -74,6 +75,10 @@
     [self onGoBack:call result:result];
   } else if ([[call method] isEqualToString:@"goForward"]) {
     [self onGoForward:call result:result];
+  } else if ([[call method] isEqualToString:@"reload"]) {
+    [self onReload:call result:result];
+  } else if ([[call method] isEqualToString:@"currentUrl"]) {
+    [self onCurrentUrl:call result:result];
   } else {
     result(FlutterMethodNotImplemented);
   }
@@ -113,6 +118,16 @@
 - (void)onGoForward:(FlutterMethodCall*)call result:(FlutterResult)result {
   [_webView goForward];
   result(nil);
+}
+
+- (void)onReload:(FlutterMethodCall*)call result:(FlutterResult)result {
+  [_webView reload];
+  result(nil);
+}
+
+- (void)onCurrentUrl:(FlutterMethodCall*)call result:(FlutterResult)result {
+  _currentUrl = [[_webView URL] absoluteString];
+  result(_currentUrl);
 }
 
 - (void)applySettings:(NSDictionary<NSString*, id>*)settings {