+++ packages/camera/CHANGELOG.md
@@ -1,7 +1,11 @@
-## 0.2.7
+## 0.2.8
 
 * Add access to the image stream from Dart.
 * Use `cameraController.startImageStream(listener)` to process the images.
+
+## 0.2.7
+
+* Fix issue with crash when the physical device's orientation is unknown.
 
 ## 0.2.6