+++ packages/webview_flutter/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.2
+
+* Added a reload method to the WebView controller.
+
+## 0.1.1
+
+* Added a `currentUrl` accessor for the WebView controller to look up what URL
+  is being displayed.
+
 ## 0.1.0+1
 
 * Fix null crash when initialUrl is unset on iOS.